adopt useUpdatePlaylistLibrary
