+++ packages/common/src/api/index.ts
@@ -82,3 +82,4 @@ export * from './tan-query/useTrendingPlaylists'
 export * from './tan-query/useTrendingUnderground'
 export * from './tan-query/useRemixes'
 export * from './tan-query/usePlaylistLibrary'
+export * from './tan-query/useUpdatePlaylistLibrary'
